+++ packages/clients/meldemichel/src/utils/createMenus.ts
@@ -1,4 +1,5 @@
 import Attributions from '@polar/plugin-attributions'
+import Filter from '@polar/plugin-filter'
 import Fullscreen from '@polar/plugin-fullscreen'
 import LayerChooser from '@polar/plugin-layer-chooser'
 import Zoom from '@polar/plugin-zoom'
@@ -11,6 +12,11 @@ export default function (mode: keyof typeof MODE): Menu[] {
       plugin: LayerChooser({}),
       icon: 'fa-book-atlas',
       id: 'layerChooser',
+    },
+    mode === MODE.COMPLETE && {
+      plugin: Filter({}),
+      icon: 'fa-filter',
+      id: 'filter',
     },
     {
       plugin: Zoom({ renderType: 'iconMenu' }),