+++ packages/plugins/Filter/src/language.ts
@@ -1,0 +1,22 @@
+import { LanguageOption } from '@polar/lib-custom-types'
+
+const language: LanguageOption[] = [
+  {
+    type: 'de',
+    resources: {
+      plugins: {
+        filter: {},
+      },
+    },
+  },
+  {
+    type: 'en',
+    resources: {
+      plugins: {
+        filter: {},
+      },
+    },
+  },
+]
+
+export default language