+++ packages/plugins/Filter/src/types.ts
@@ -1,0 +1,3 @@
+export interface FilterState {}
+
+export type FilterGetters = FilterState